+++ presentation/src/main/java/com/example/huitdduru/view/main/MainActivity.kt
@@ -3,7 +3,7 @@ package com.example.huitdduru.view.main
 import android.os.Bundle
 import com.example.huitdduru.R
 import com.example.huitdduru.databinding.ActivityMainBinding
-import com.example.huitdduru.view.base.BaseActivity
+import com.example.huitdduru.base.BaseActivity
 
 class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
     override fun onCreate(savedInstanceState: Bundle?) {
