+++ presentation/src/main/java/com/example/huitdduru/view/sign/CertifyFragment.kt
@@ -7,7 +7,7 @@ import com.example.huitdduru.R
 import com.example.huitdduru.databinding.FragmentCertifyBinding
 import com.example.huitdduru.util.ToastType
 import com.example.huitdduru.util.repeatOnStarted
-import com.example.huitdduru.view.base.BaseFragment
+import com.example.huitdduru.base.BaseFragment
 import com.example.huitdduru.viewmodel.register.RegisterViewModel
 import com.example.huitdduru.viewmodel.register.RegisterViewModel.Event
 import dagger.hilt.android.AndroidEntryPoint
