+++ presentation/src/main/java/com/example/huitdduru/view/sign/LoginActivity.kt
@@ -4,7 +4,7 @@ import android.content.Intent
 import android.os.Bundle
 import com.example.huitdduru.R
 import com.example.huitdduru.databinding.ActivityLoginBinding
-import com.example.huitdduru.view.base.BaseActivity
+import com.example.huitdduru.base.BaseActivity
 
 class LoginActivity : BaseActivity<ActivityLoginBinding>(R.layout.activity_login) {
     override fun onCreate(savedInstanceState: Bundle?) {
