+++ presentation/src/main/java/com/example/huitdduru/view/sign/SuccessActivity.kt
@@ -3,7 +3,7 @@ package com.example.huitdduru.view.sign
 import android.os.Bundle
 import com.example.huitdduru.R
 import com.example.huitdduru.databinding.ActivitySuccessBinding
-import com.example.huitdduru.view.base.BaseActivity
+import com.example.huitdduru.base.BaseActivity
 
 class SuccessActivity : BaseActivity<ActivitySuccessBinding>(R.layout.activity_success) {
     override fun onCreate(savedInstanceState: Bundle?) {
